rename 3d detector
